tested mongoRDD.count() run on large dataset
