+++ pom.xml
@@ -12,10 +12,25 @@
     <properties>
         <lucene.version>6.2.0</lucene.version>
         <spark.hdp.version>1.6.0.2.4.1.1-3</spark.hdp.version>
-        <scala.version>2.10.5</scala.version>
-        <scala.binary.version>2.10</scala.binary.version>
+        <scala.version>2.11.8</scala.version>
+        <scala.binary.version>2.11</scala.binary.version>
     </properties>
-
+    <repositories>
+        <repository>
+            <id>central</id>
+            <url>https://repo1.maven.org/maven</url>
+            <releases>
+                <enabled>true</enabled>
+            </releases>
+        </repository>
+        <repository>
+            <id>central2</id>
+            <url>https://repo1.maven.org/maven2</url>
+            <releases>
+                <enabled>true</enabled>
+            </releases>
+        </repository>
+    </repositories>
     <dependencies>
         <!-- https://mvnrepository.com/artifact/org.scala-lang/scala-reflect -->
 
@@ -60,8 +75,8 @@
 
         <dependency>
             <groupId>org.apache.spark</groupId>
-            <artifactId>spark-core_2.10</artifactId>
-            <version>1.6.1</version>
+            <artifactId>spark-core_2.11</artifactId>
+            <version>2.0.1</version>
         </dependency>
        <!-- <dependency>
             <groupId>org.apache.spark</groupId>
@@ -79,9 +94,9 @@
         </dependency>
  -->
         <dependency>
-            <groupId>org.mongodb</groupId>
+            <groupId>org.mongodb.mongo-hadoop</groupId>
             <artifactId>mongo-hadoop-core</artifactId>
-            <version>1.3.0</version>
+            <version>1.5.1</version>
         </dependency>
         <dependency>
             <groupId>org.elasticsearch</groupId>
@@ -99,11 +114,11 @@
             <artifactId>hadoop-client</artifactId>
             <version>2.3.0</version>
         </dependency>
-       <!-- <dependency>
+       <dependency>
             <groupId>org.mongodb.spark</groupId>
             <artifactId>mongo-spark-connector_2.10</artifactId>
             <version>2.0.0</version>
-        </dependency> -->
+        </dependency>
        <!-- <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-sql_2.10</artifactId>
@@ -113,14 +128,21 @@
         <dependency>
             <groupId> org.mongodb </groupId>
             <artifactId> mongo-java-driver </artifactId>
-            <version>  2.10.0</version>
+            <version>  3.2.1</version>
         </dependency>
 
+        <!--<dependency>
+            <groupId>org.mongodb</groupId>
+            <artifactId>bson</artifactId>
+            <version>2.3</version>
+        </dependency> -->
+        <!-- https://mvnrepository.com/artifact/org.mongodb/bson -->
         <dependency>
             <groupId>org.mongodb</groupId>
             <artifactId>bson</artifactId>
             <version>2.3</version>
         </dependency>
+
         <!-- https://mvnrepository.com/artifact/org.apache.spark/spark-mllib_2.10 -->
      <!--  <dependency>
             <groupId>org.apache.spark</groupId>
